*** empty log message ***
git-svn-id: https://svn.code.sf.net/p/confix/svn/confix/trunk@709 ff2b13fb-00b2-4ffb-9011-bdd21bf2a437

+++ libconfix/testutils/persistent.py
@@ -40,7 +40,9 @@ class PersistentTestCase(unittest.TestCase):
     def tearDown(self):
         dir = os.sep.join(self.__my_rootpath)
         if os.path.isdir(dir):
-            shutil.rmtree(dir)
+            if os.environ.get('KEEP_PERSISTENT_TEST') is None:
+                shutil.rmtree(dir)
+                pass
             pass
         pass
